Split execution into stages

+++ cli.ts
@@ -1,34 +1,20 @@
-// Config
-// Run with all - to execute both llm and exec
-// Run with llm - to execute only llm steps
-// Run with exec - to execute only exec steps
+
+// CLI Config
 export async function getConfig(): Promise<
-  { run_llm: boolean; run_exec: boolean }
+  { run_llm: boolean; run_exec: boolean; run_shinkai: boolean }
 > {
-  const mode = Deno.args[0];
-  let run_llm = false;
-  let run_exec = false;
-  switch (mode) {
-    case "exec":
-      run_exec = true;
-      break;
-    case "llm":
-      try {
-        if (await Deno.stat("./results")) {
-          await Deno.remove("./results", { recursive: true });
-        }
-      } catch (_) { /* nop */ }
-      run_llm = true;
-      break;
-    case "all":
-    default:
-      try {
-        if (await Deno.stat("./results")) {
-          await Deno.remove("./results", { recursive: true });
-        }
-      } catch (_) { /* nop */ }
-      run_llm = true;
-      run_exec = true;
+  const args = Deno.args;
+
+  const run_llm = args.includes("llm") || args.length === 0;
+  const run_exec = args.includes("exec") || args.length === 0;
+  const run_shinkai = args.includes("shinkai") || args.length === 0;
+
+  if (run_shinkai) {
+    try {
+      if (await Deno.stat("./results")) {
+        await Deno.remove("./results", { recursive: true });
+      }
+    } catch (_) { /* nop */ }
   }
-  return { run_llm, run_exec };
+  return { run_llm, run_exec, run_shinkai };
 }
